requirements.txt

Django==4.1.5
django-crispy-forms==1.14.0
dj-database-url
psycopg2-binary
whitenoise[brotli]
gunicorn
djangorestframework